fix(vue-renderer): Move logs to vue-renderer to get plugins logs
